+++ frontend/app/components/Divider.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import React from "react";
+
+const Divider: React.FC = () => {
+  return <div className="h-px w-full bg-gray-200"></div>;
+};
+
+export default Divider;
